chore: move atan2 test
